add function to make backup

+++ setup.sh
@@ -25,8 +25,13 @@ for file in $(ls -A $core_dir); do
   # get basename
   filename=$(basename $file)
   
-  # skip if symbolic link exists
-  [[ -f ~/$filename ]] && showmsg "~/$filename exists" && continue
+  # if symbolic link exists
+  if [[ -f ~/$filename ]]; then
+    read -p "~/$filename exists. replace? (Y/n)" ans
+    [[ $ans != 'Y' ]] && continue
+    # create backup
+    mv ~/$filename{,.$(date +%Y%m%d)}
+  fi
 
   ln -s "${core_dir}/$file" ~/$filename
   showmsg "created symbolic link ~/$filename"
